calc fix

+++ src/main/java/ru/zhidkova/calcapp/util/Calculator.java
@@ -1,16 +1,16 @@
 package ru.zhidkova.calcapp.util;
 
 public class Calculator {
-    public static double sum(double num1, double num2) {
-        double sumResult = num1 + num2;
+    public static int sum(int num1, int num2) {
+        int sumResult = num1 + num2;
         return sumResult;
     };
-    public static double subtract(double num1, double num2) {
-        double substractResult = num1 - num2;
+    public static int subtract(int num1, int num2) {
+        int substractResult = num1 - num2;
         return substractResult;
     };
-    public static double multiply(double num1, double num2) {
-        double multiplyResult = num1 * num2;
+    public static int multiply(int num1, int num2) {
+        int multiplyResult = num1 * num2;
         return multiplyResult;
     };
 }
